+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - prevent accidental task sidebar resize on click
 - added drag and drop support for tasks and subtasks
 - avoid showing password dialog on start up in some cases
+- added getOpenProjects to Extension Context API
 
 ## [0.75.0]